Avoid::BoardServiceImpl

+++ src/main/java/com/project/yjshop/service/board/BoardServiceImpl.java
@@ -3,15 +3,22 @@ package com.project.yjshop.service.board;
 import com.project.yjshop.domain.board.Board;
 import com.project.yjshop.domain.board.BoardRepository;
 import com.project.yjshop.domain.image.ImageRepository;
+import com.project.yjshop.error.ErrorCode;
+import com.project.yjshop.error.exception.CustomException;
+import com.project.yjshop.security.auth.PrincipalDetails;
 import com.project.yjshop.service.image.ImageServiceImpl;
 import com.project.yjshop.service.image.S3Service;
 import com.project.yjshop.web.payload.request.board.ProductRequest;
 import com.project.yjshop.web.payload.response.board.ProductResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
 
 @RequiredArgsConstructor
 @Service
@@ -23,28 +30,49 @@ public class BoardServiceImpl implements BoardService{
     private final S3Service s3Service;
 
     @Override
-    public ProductResponse posting(ProductRequest productRequest) throws IOException {
+    public ProductResponse posting(ProductRequest productRequest,
+                                   BindingResult bindingResult,
+                                   PrincipalDetails principalDetails) throws IOException {
 
-        Board saveBoard = boardRepository.save(Board.builder()
-                .title(productRequest.getTitle())
-                .titleImage(imageRepository.findByImagePath(imageService.imageUpload(s3Service.upload(productRequest.getTitleImage()))).get())
-                .price(productRequest.getPrice())
-                .build());
+        if (bindingResult.hasErrors()) {
+            Map<String, String> errorMap = new HashMap<>();
 
-        return ProductResponse.builder()
-                .message("Post success")
-                .product(ProductResponse.Product.builder()
-                        .boardId(saveBoard.getId())
-                        .title(saveBoard.getTitle())
-                        .titleImage(saveBoard.getTitleImage().getImageFullPath())
-                        .price(saveBoard.getPrice())
-                        .build())
-                .build();
+            for (FieldError error : bindingResult.getFieldErrors()) {
+                errorMap.put(error.getField(), error.getDefaultMessage());
+            }
+            throw new CustomException(ErrorCode.POSTING_FAILED, errorMap);
+        } else {
+
+            Board saveBoard = boardRepository.save(Board.builder()
+                    .title(productRequest.getTitle())
+                    .titleImage(imageRepository.findByImagePath(imageService.imageUpload(s3Service.upload(productRequest.getTitleImage()))).get())
+                    .price(productRequest.getPrice())
+                    .count(productRequest.getCount())
+                    .user(principalDetails.getUser())
+                    .build());
+
+            return ProductResponse.builder()
+                    .message("Post success")
+                    .product(ProductResponse.Product.builder()
+                            .boardId(saveBoard.getId())
+                            .title(saveBoard.getTitle())
+                            .titleImage(saveBoard.getTitleImage().getImageFullPath())
+                            .price(saveBoard.getPrice())
+                            .count(saveBoard.getCount())
+                            .build())
+                    .build();
+        }
     }
 
     @Override
-    public ProductResponse deleting(Long boardId) {
+    public ProductResponse deleting(Long boardId, PrincipalDetails principalDetails) {
         Board delBoard = boardRepository.findById(boardId).get();
+
+        if(principalDetails.getUser().getId() != delBoard.getUser().getId()) {
+            throw new CustomException(ErrorCode.USER_NOT_MATCH);
+        }
+
+
         s3Service.delete(delBoard.getTitleImage().getImagePath());
         boardRepository.delete(delBoard);
 
@@ -55,6 +83,7 @@ public class BoardServiceImpl implements BoardService{
                         .title(delBoard.getTitle())
                         .titleImage(delBoard.getTitleImage().getImageFullPath())
                         .price(delBoard.getPrice())
+                        .count(delBoard.getCount())
                         .build())
                 .build();
     }
